feat: pressing enter to select an emoji
